Resoluções finail

+++ src/services/transactions/index.ts
@@ -1,55 +1,58 @@
-import { ITransaction } from "@/types/transaction";
+import { ITransaction, ITransactionResponse } from "@/types/transaction";
 import { api } from "../api";
 import { toast } from "react-toastify";
 import { QueryFunctionContext } from "@tanstack/react-query";
 
 export async function getTransactions(
   ctx: QueryFunctionContext<readonly [string, number, number]>
-) {
+): Promise<ITransactionResponse> {
   const [_key, skip, take] = ctx.queryKey;
 
   const response = await api.get("/transaction", {
     params: { skip, take },
   });
 
-  return {
-    transactions: response.data.transactions,
-    totalCount: response.data.totalCount,
-  };
+  return response.data;
 }
-
 export async function createTransaction(transaction: ITransaction) {
   try {
     const response = await api.post("/transaction", transaction);
     toast.success("Transação adicionada com sucesso!");
     return response.data;
   } catch (error) {
-    throw new Error("Erro ao criar transação: " + error);
+    toast.error("Erro ao criar transação.");
   }
 }
 
-export async function uptadeTransaction({
+export async function updateTransaction({
   id,
   data,
 }: {
   id: string | undefined;
   data: Partial<ITransaction>;
 }) {
+  if (!id) {
+    toast.error("ID da transação não encontrado.");
+    return;
+  }
   try {
     const response = await api.patch(`/transaction/${id}`, data);
     toast.success("Transação atualizada com sucesso");
     return response.data;
   } catch (error) {
     toast.error("Erro ao atualizar transação");
-    throw error;
   }
 }
 
 export async function deleteTransaction(id: string | undefined) {
+  if (!id) {
+    toast.error("ID da transação não encontrado.");
+    return;
+  }
   try {
     await api.delete(`/transaction/${id}`);
     toast.success("Transação deletada com sucesso!");
   } catch (error) {
-    throw new Error("Erro ao deletar a transação: " + error);
+    toast.error("Erro ao deletar a transação.");
   }
 }
